Agregue esquema e instancia de base de datos

+++ InterfazServer.java
@@ -2,7 +2,7 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface InterfazServer extends Remote {
-    float nuevaVenta(InterfazVenta venta, String vendedor) throws RemoteException;
+    int nuevaVenta(InterfazVenta venta, String vendedor) throws RemoteException;
     String[] listarArticulos() throws RemoteException;
     Venta[] listarVentas()throws RemoteException;
 	boolean esArticuloValido(String nombreArticulo) throws RemoteException;
